Renaming style imports.
