+++ components/button/button-subtle.js
@@ -1,9 +1,9 @@
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { LitElement, html } from 'lit-element/lit-element.js';
-import { label } from '../typography/styles.js';
-import { buttonSubtle } from './button-subtle-styles.js';
-import { button } from './styles.js';
-import { D2LButtonMixin } from './mixin.js';
+import { D2LButtonMixin } from './button-mixin.js';
+import { labelStyles } from '../typography/styles.js';
+import { buttonSharedStyles } from './button-shared-styles.js';
+import { buttonSubtleStyles } from './button-subtle-styles.js';
 import 'd2l-icons/d2l-icon.js';
 import 'd2l-icons/tier1-icons.js';
 
@@ -23,7 +23,7 @@ export class D2LButtonSubtle extends D2LButtonMixin(LitElement) {
 	}
 
 	static get styles() {
-		return [ label, button, buttonSubtle ];
+		return [ labelStyles, buttonSharedStyles, buttonSubtleStyles ];
 	}
 
 	constructor() {
